+++ packages/xo-server-test/src/issues/index.spec.js
@@ -1,5 +1,7 @@
 /* eslint-env jest */
 
+import { parseDuration } from '@vates/parse-duration'
+
 import config from '../_config'
 import xo from '../_xoConnection'
 
@@ -47,5 +49,70 @@ describe('issue', () => {
     expect(typeof id).toBe('string')
 
     await xo.call('network.delete', { id })
+  })
+
+  describe('4980', () => {
+    let template
+    beforeAll(async () => {
+      jest.setTimeout(parseDuration(config.cloneTempVmTimeout))
+      template = await xo.cloneTempVm(config.templates.default)
+    })
+
+    const bootOrder = 'cd'
+    const virtualizationMode = 'hvm'
+    beforeAll(async () => {
+      await Promise.all([
+        xo.call('vm.set', {
+          id: template.id,
+          virtualizationMode,
+        }),
+        xo.call('vm.setBootOrder', { vm: template.id, order: bootOrder }),
+      ])
+      await xo.waitObjectState(template.id, {
+        virtualizationMode,
+        boot: {
+          order: bootOrder,
+        },
+      })
+    })
+
+    test('create vm with disks should keep the template boot order', async () => {
+      const vm = await xo.createTempVm({
+        template: template.id,
+        VDIs: [
+          {
+            size: 1,
+            SR: config.srs.default,
+            type: 'user',
+          },
+        ],
+      })
+      expect(vm.boot.order).toBe(bootOrder)
+    })
+
+    test('create vm without disks should make network boot the first option', async () => {
+      const vm = await xo.createTempVm({
+        template: template.id,
+      })
+      expect(vm.boot.order).toBe('n' + bootOrder)
+    })
+
+    test('create vm with disks and network installation should make network boot the first option', async () => {
+      const vm = await xo.createTempVm({
+        template: template.id,
+        installation: {
+          method: 'network',
+          repository: 'pxe',
+        },
+        VDIs: [
+          {
+            size: 1,
+            SR: config.srs.default,
+            type: 'user',
+          },
+        ],
+      })
+      expect(vm.boot.order).toBe('n' + bootOrder)
+    })
   })
 })