+++ packages/xo-web/src/common/xo/index.js
@@ -1,14 +1,13 @@
 import asap from 'asap'
-import cookies from 'cookies-js'
+import cookies from 'js-cookie'
 import fpSortBy from 'lodash/fp/sortBy'
-import pFinally from 'promise-toolbox/finally'
 import React from 'react'
-import reflect from 'promise-toolbox/reflect'
-import tap from 'promise-toolbox/tap'
 import updater from 'xoa-updater'
 import URL from 'url-parse'
 import Xo from 'xo-lib'
 import { createBackoff } from 'jsonrpc-websocket-client'
+import { get as getDefined } from '@xen-orchestra/defined'
+import { pFinally, reflect, tap, tapCatch } from 'promise-toolbox'
 import { SelectHost } from 'select-objects'
 import {
   filter,
@@ -73,7 +72,7 @@ export const signOut = () => {
   // prevent automatic reconnection
   xo.removeListener('closed', connect)
 
-  cookies.expire('token')
+  cookies.remove('token')
   window.location.reload(true)
 }
 
@@ -204,7 +203,10 @@ const createSubscription = cb => {
     timeout = setTimeout(clearCache, clearCacheDelay)
   }
 
-  const loop = () => {
+  // will loop if n > 0 at the end
+  //
+  // will not do anything if already running
+  const run = () => {
     clearTimeout(timeout)
 
     if (running) {
@@ -222,7 +224,7 @@ const createSubscription = cb => {
             return uninstall()
           }
 
-          timeout = setTimeout(loop, delay)
+          timeout = setTimeout(run, delay)
 
           if (!isEqual(result, cache)) {
             cache = result
@@ -258,7 +260,7 @@ const createSubscription = cb => {
     }
 
     if (n++ === 0) {
-      loop()
+      run()
     }
 
     return once(() => {
@@ -272,7 +274,7 @@ const createSubscription = cb => {
 
   subscribe.forceRefresh = () => {
     if (n) {
-      loop()
+      run()
     }
   }
 
@@ -373,7 +375,7 @@ const setNotificationCookie = (id, changes) => {
   cookies.set(
     `notifications:${store.getState().user.id}`,
     JSON.stringify(notifications),
-    { expires: Infinity }
+    { expires: 9999 }
   )
 }
 
@@ -564,9 +566,13 @@ export const editServer = (server, props) =>
   )
 
 export const enableServer = server =>
-  _call('server.enable', { id: resolveId(server) })::pFinally(
-    subscribeServers.forceRefresh
-  )
+  _call('server.enable', { id: resolveId(server) })
+    ::tapCatch(error => {
+      if (error.message === 'Invalid XML-RPC message') {
+        error(_('enableServerErrorTitle'), _('enableServerErrorMessage'))
+      }
+    })
+    ::pFinally(subscribeServers.forceRefresh)
 
 export const disableServer = server =>
   _call('server.disable', { id: resolveId(server) })::tap(
@@ -1182,7 +1188,10 @@ export const copyVms = (vms, type) => {
   }, noop)
 }
 
-export const copyVm = vm => copyVms([vm])
+export const copyVm = async vm => {
+  const result = await copyVms([vm], vm.type)
+  return getDefined(() => result[0])
+}
 
 export const convertVmToTemplate = vm =>
   confirm({
@@ -1774,15 +1783,59 @@ export const deleteVifs = vifs =>
 
 export const setVif = (
   vif,
-  { allowedIpv4Addresses, allowedIpv6Addresses, mac, network, rateLimit }
+  {
+    allowedIpv4Addresses,
+    allowedIpv6Addresses,
+    lockingMode,
+    mac,
+    network,
+    rateLimit,
+    txChecksumming,
+  }
 ) =>
   _call('vif.set', {
     allowedIpv4Addresses,
     allowedIpv6Addresses,
     id: resolveId(vif),
+    lockingMode,
     mac,
     network: resolveId(network),
     rateLimit,
+    txChecksumming,
+  })
+
+export const getLockingModeValues = () => _call('vif.getLockingModeValues')
+
+export const addAclRule = ({
+  allow,
+  protocol = undefined,
+  port = undefined,
+  ipRange = '',
+  direction,
+  vif,
+}) =>
+  _call('sdnController.addRule', {
+    allow,
+    protocol,
+    port,
+    ipRange,
+    direction,
+    vifId: resolveId(vif),
+  })
+
+export const deleteAclRule = ({
+  protocol = undefined,
+  port = undefined,
+  ipRange = '',
+  direction,
+  vif,
+}) =>
+  _call('sdnController.deleteRule', {
+    protocol,
+    port,
+    ipRange,
+    direction,
+    vifId: resolveId(vif),
   })
 
 // Network -----------------------------------------------------------
@@ -3180,15 +3233,39 @@ export const destroyProxyAppliances = proxies =>
 export const upgradeProxyAppliance = proxy =>
   _call('proxy.upgradeAppliance', { id: resolveId(proxy) })
 
-export const checkProxyHealth = proxy =>
-  _call('proxy.checkHealth', { id: resolveId(proxy) }).then(() =>
-    success(
-      <span>
-        <Icon icon='success' /> {_('proxyTestSuccess', { name: proxy.name })}
-      </span>,
-      _('proxyTestSuccessMessage')
-    )
-  )
+export const getProxyApplianceUpdaterState = id =>
+  _call('proxy.getApplianceUpdaterState', { id })
+
+const PROXY_HEALTH_CHECK_COMMON_ERRORS_CODE = new Set([
+  'ECONNREFUSED',
+  'ECONNRESET',
+  'EHOSTUNREACH',
+  'ENOTFOUND',
+  'ETIMEDOUT',
+])
+
+export const checkProxyHealth = async proxy => {
+  const result = await _call('proxy.checkHealth', {
+    id: resolveId(proxy),
+  })
+  return result.success
+    ? success(
+        <span>
+          <Icon icon='success' /> {_('proxyTestSuccess', { name: proxy.name })}
+        </span>,
+        _('proxyTestSuccessMessage')
+      )
+    : error(
+        <span>
+          <Icon icon='error' /> {_('proxyTestFailed', { name: proxy.name })}
+        </span>,
+        <span>
+          {PROXY_HEALTH_CHECK_COMMON_ERRORS_CODE.has(result.error.code)
+            ? _('proxyTestFailedConnectionIssueMessage')
+            : result.error.message}
+        </span>
+      )
+}
 
 // Audit plugin ---------------------------------------------------------